service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
gradio==3.50.2
diffusers>=0.23.1
transformers>=4.38.0
accelerate>=0.25.0
safetensors>=0.4.0
torch>=2.0.0
numpy~=1.24.0
Pillow>=10.2.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.24.1 